src/backend/services/auth/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
redis==5.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-saml==1.16.0
python-ldap==3.4.3
cryptography==41.0.7